대쉬보드 화면, 트래킹 화면 연결

+++ app/src/main/java/com/jingom/deepworktracker/feature_tracking/data/local/DeepWorkDao.kt
@@ -17,7 +17,7 @@ interface DeepWorkDao {
 	suspend fun getDeepWorkById(id: Int): DeepWork?
 
 	@Insert(onConflict = OnConflictStrategy.REPLACE)
-	suspend fun insertDeepWork(deepWork: DeepWork): Int
+	suspend fun insertDeepWork(deepWork: DeepWork): Long
 
 	@Insert(onConflict = OnConflictStrategy.REPLACE)
 	suspend fun insertDeepWorks(deepWorks: List<DeepWork>)
