+++ app/src/main/java/com/jingom/deepworktracker/feature_tracking/data/local/DeepWorkLocalDataSourceImpl.kt
@@ -27,7 +27,7 @@ class DeepWorkLocalDataSourceImpl @Inject constructor(
 		}
 	}
 
-	override suspend fun insertDeepWork(deepWork: DeepWork): Int {
+	override suspend fun insertDeepWork(deepWork: DeepWork): Long {
 		return withContext(ioDispatcher) {
 			deepWorkDao.insertDeepWork(deepWork)
 		}
