+++ app/src/main/java/com/jingom/deepworktracker/feature_tracking/data/repository/DeepWorkRepositoryImpl.kt
@@ -29,7 +29,7 @@ class DeepWorkRepositoryImpl @Inject constructor(
 		return deepWorkLocalDataSource.getDeepWorkById(id)
 	}
 
-	override suspend fun insertDeepWork(deepWork: DeepWork): Int {
+	override suspend fun insertDeepWork(deepWork: DeepWork): Long {
 		return deepWorkLocalDataSource.insertDeepWork(deepWork)
 	}
